requirements.txt
python-telegram-bot[job-queue]
google-generativeai
psycopg2-binary